requirements.txt
livekit
livekit-agents[openai,deepgram,cartesia,silero,turn_detector]
livekit-plugins-noise-cancellation
python-dotenv
vapi-python
fastapi
uvicorn
pydantic
requests
flask
